Change the Live Test multi-label classification
I've changed way in which the "Live Test" tool selects the categories.
The previous method was ad hoc and didn't produce good results in
general. For the one version, I decided to implement the K-Means
algorithm to divide the categories into two clusters, one for
"select as classified" and another for "ignore". The key idea is to
partitioned the categories according to the obtained confidence value.