Change the way unknown categories are handled
Previously, ``get_category_name`` or ``get_category_index`` threw
``InvalidCategory`` exception when called with a non-existing category
label or index, respectively. Now, instead, they'll return the
constants STR_UNKNOWN_CATEGORY or IDX_UNKNOWN_CATEGORY, respectively.
Methods that depended on these methods to throw this exception have
been changed to check manually throw it when it is needed.

This changes were made with the aim to simplify the code of methods
like ``predict``, ``classify_label`` or ``classify_multilabel``, when
called with ``def_cat=``unknown``. Before, "unknown" category will be
assigned the index len(categories), but only inside this methods, which
semantically it is not so consistent. We changed ``get_category_name``
and ``get_category_index`` to return themselves always the same index
to literally any unknown category (IDX_UNKNOWN_CATEGORY), independently
from where it is called from or the number of categories, which is more
consistent and semantically clearer.