Add ``extract_insight`` method to SS3 class
This new method, given a document, returns the pieces of text that
were involved in the classification decision, along with the conf-
idence values associated with them. Read the documentation for mo-
re info.

This method was created to provide the user with a different way
of checking  what SS3 is actually  learning. Before this  method
the user had to use the "Live Test" tool  to do it. Furthermore,
this method could be used to "summary" the input document, leav-
ing only relevant parts to whom it needs them.