Improve how text is (re.)split and reconstructed
I've changed the way by which the document was split when calling the
``classify`` method with ``json=True``. Now the split is done in such a
way so that the JSON output contains exactly all the text needed to
reconstruct the exact same input document just by joining each lexeme
with the empty string. i.e. using ``"".join(...)``. Additionally, I had
to add an extra condition inside `__classify_sentence__` to allow SS3
to still recognize word n-grams even when word delimiters are kept
in-between these words.

This improvement allow simplifying the reconstruction process of the
original input document for modules and applications that require such a
feature (e.g. the `extract_insight` method, the Live Test tool). Before
this improvement, those using this JSON output, had to use ad hoc
methods to try to reconstruct the input document, which never produced
the exact same original document.