Improve `predict` speed (10x-200x faster!)
The `predict` method (and therefore all methods using it, such as
`grid_search`) now runs form 10 to (above) 200 times faster (e.g. from 2
mins to 9s, or even 500ms when used the same `x_test` a second time,
which is the usual case in `grid_search(es)`).

Now, the only bottleneck left to speed up the overall `grid_search` time
is training time, which is something in "the TODO list" (i.e. rewrite
the training algorithm to use numpy and sparse matrices)