Improve word n-grams recognition at classification
The method used to recognize/parse learned word n-grams was improved.
The new code it is easier to follow and clearer, and semantically more
accurate. Furthermore, single words and word n-grams tokanization and
visualization in the Live Test Tool was also improved by this new code.

Before this modification, there was a bug in the Live Test tool raised
by the addition of the custom preprocessing support (#3). Some words
were visualized in the wrong way, especially those ended with question
marks, for instance the sentence "Great  atmosphere then ?????". In
addition, now the user can put an arbitrarily number of spaces between
words and they will be ignored when recognizing word n-grams. For
instance, before this modification, if the user entered:
"machine     learning"
It was not recognized as a bigram due to those extra spaces.