Add preprocessing argument to Evaluation methods
The ``Evaluation`` class didn't include the "prep" argument to
disable/enable the default preprocessing method. As a consequence, users
were not be able to perform any hyperparameter optimization o model
evaluation using only their custom preprocessing method. Therefore, the
"prep" argument was added to the ``grid_search()`, ``test()``, and
``kfold_cross_validation()`` functions of the ``Evaluation`` class.

Plus, the tutorial notebook for custom preprocessing was updated to
include an example of how to use the ``Evaluation.grid_search()`` to
find the best hyperparameter values with the new "prep" argument.