Add multilabel support to train/fit methods (#5)
The fit/train method now supports multilabel classification. It will
automatically determine if we're dealing with a multilabel
classification problem by looking at the first item of the `y_train`
list. If the first item is a list (of labels), i.e., if it's not a
single label, it will assume we're dealing with a multilabel
classification problem.