Change multilabel classification default category
``predict`` and ``classify_multilabel`` now have "unknown" as the
default category when working with multi-label classification. When this
category is used it will return an empty list [] to indicate that the
predicted category is unknown.

In addition, ``kmean_multilabel_size`` returns 0 when the centroid of
the two clusters have the same value, which means that ``predict`` or
``classify_multilabel`` will return an empty list [] when predicted
confidence values are all equal.