Fix little bug in Live Test
There was a little bug the the Live Test Tool when working with
multi-label classification. When the true label's index was 0 it made
the ``!cat_info`` condition true in the
``$scope.is_cat_active(cat_info)`` function, returning ``false``, which
made the true label turn red (as if it were misclassified).

Solution: simple! replaced !cat_info with cat_info == null in the if
statement.