Fix failing test for server
Due to the last committed change, now the "live_test" command always
assigns the current model to the Live Test Server. Therefore, the server
test "get_info" request now returns the correct model name ("cmd_test"),
instead of the default name ("model"), so the assert checking for the
model name was removed since it is not really necessary.