Add word clouds generation capability
The method "save_wordcloud" has been added to the SS3 main class. This
method allows easily generating word clouds for the given category. The
user can customize the cloud through its different arguments, for
instance, set the number of words to be displayed, the background and
text color, and whether to use words, bigrams, trigrams, or any other
custom n-gram to construct the could.