Add SS3Vectorizer class to pyss3 module
The "SS3Vectorizer" class has been added to the pyss3 main module. This
class was inspired by the ``CountVectorizer`` and ``TfidfVectorizer`` of
the Scikit-learn's ``sklearn.feature_extraction.text`` submodule. This
submodule gathers utilities to build feature vectors from text
documents. Thus, the SS3Vectorizer class converts a collection of raw
documents to a document-term matrix of SS3-based features. That is, this
class allows vectorizing documents using SS3 as a weighting scheme (in
different ways) to use them with any other model.